Refactor view/com/post/Post and remove temporary 2 components
